Refactor field page objects in cypress
