+++ tests/cypress/page-object/fields/field.ts
@@ -1,0 +1,6 @@
+import {BaseComponent} from "@jahia/cypress";
+
+export class Field extends BaseComponent {
+    fieldName: string;
+    multiple: boolean;
+}
