+++ tests/cypress/page-object/fields/index.js
@@ -1,0 +1,4 @@
+export * from './field'
+export * from './pickerField';
+export * from './richTextField';
+export * from './smallTextField';
