+++ tests/cypress/page-object/fields/richTextField.ts
@@ -1,8 +1,8 @@
 import {BaseComponent, getComponentByRole} from '@jahia/cypress';
-import {Picker} from './picker';
+import {Picker} from '../picker';
+import {Field} from "./field";
 
-export class RichTextField extends BaseComponent {
-    fieldName: string;
+export class RichTextField extends Field {
 
     type(text) {
         this.get().iframe('.cke_wysiwyg_frame').type(text);
